requirements.txt
alembic==1.16.4
blinker==1.9.0
boto3==1.40.21
botocore==1.40.21
click==8.2.1
Flask==3.1.2
flask-cors==6.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
graphql-core==3.2.6
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
lia-web==0.2.3
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
s3transfer==0.13.1
six==1.17.0
SQLAlchemy==2.0.43
strawberry-graphql==0.280.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
